fix: Game: moved tileColor() to correct translation unit
Created tile.cpp for tileColor() and updated CMakeLists.txt and
meson.build accordingly.
