+++ src/tile.cpp
@@ -1,0 +1,12 @@
+#include "tile.hpp"
+#include <cmath>
+#include <vector>
+
+Color::Modifier Tile::tileColor(ull value) {
+  std::vector<Color::Modifier> colors{red, yellow, magenta, blue, cyan, yellow,
+                                      red, yellow, magenta, blue, green};
+  int log = log2(value);
+  int index = log < 12 ? log - 1 : 10;
+
+  return colors[index];
+}
